hf_space/requirements.txt
gitpython==3.1.40
paramiko==3.3.1
python-dotenv==1.0.0
matplotlib==3.8.1
numpy==1.24.1
scipy==1.11.3
scikit-learn==1.3.2
